requirements.txt
Pillow==7.2.0
tqdm==4.46.0
requests==2.23.0